First trial to implement broadcast_arguments functionality as a mixin
