+++ qiskit/circuit/barrier.py
@@ -15,9 +15,10 @@
 from qiskit.exceptions import QiskitError
 from .instruction import Instruction
 from .operation import Operation
+from .argumentsbroadcaster import ArgumentsBroadcasterBarrier
 
 
-class Barrier(Instruction, Operation):
+class Barrier(ArgumentsBroadcasterBarrier, Instruction, Operation):
     """Barrier instruction."""
 
     _directive = True
@@ -29,9 +30,6 @@ class Barrier(Instruction, Operation):
     def inverse(self):
         """Special case. Return self."""
         return Barrier(self.num_qubits)
-
-    def broadcast_arguments(self, qargs, cargs):
-        yield [qarg for sublist in qargs for qarg in sublist], []
 
     def c_if(self, classical, val):
         raise QiskitError("Barriers are compiler directives and cannot be conditional.")
