+++ qiskit/circuit/delay.py
@@ -17,9 +17,9 @@ import numpy as np
 from qiskit.circuit.exceptions import CircuitError
 from qiskit.circuit.instruction import Instruction
 from qiskit.circuit.parameterexpression import ParameterExpression
+from .argumentsbroadcaster import ArgumentsBroadcasterDelay
 
-
-class Delay(Instruction):
+class Delay(ArgumentsBroadcasterDelay, Instruction):
     """Do nothing and just delay/wait/idle for a specified duration."""
 
     def __init__(self, duration, unit="dt"):
@@ -32,9 +32,6 @@ class Delay(Instruction):
     def inverse(self):
         """Special case. Return self."""
         return self
-
-    def broadcast_arguments(self, qargs, cargs):
-        yield [qarg for sublist in qargs for qarg in sublist], []
 
     def c_if(self, classical, val):
         raise CircuitError("Conditional Delay is not yet implemented.")
