+++ qiskit/circuit/gate.py
@@ -20,9 +20,10 @@ from qiskit.circuit.parameterexpression import ParameterExpression
 from qiskit.circuit.exceptions import CircuitError
 from .instruction import Instruction
 from .operation import Operation
+from .argumentsbroadcaster import ArgumentsBroadcasterGate
 
 
-class Gate(Instruction, Operation):
+class Gate(ArgumentsBroadcasterGate, Instruction, Operation):
     """Unitary gate."""
 
     def __init__(
@@ -118,102 +119,6 @@ class Gate(Instruction, Operation):
         from .add_control import add_control
 
         return add_control(self, num_ctrl_qubits, label, ctrl_state)
-
-    @staticmethod
-    def _broadcast_single_argument(qarg: List) -> List:
-        """Expands a single argument.
-
-        For example: [q[0], q[1]] -> [q[0]], [q[1]]
-        """
-        # [q[0], q[1]] -> [q[0]]
-        #              -> [q[1]]
-        for arg0 in qarg:
-            yield [arg0], []
-
-    @staticmethod
-    def _broadcast_2_arguments(qarg0: List, qarg1: List) -> List:
-        if len(qarg0) == len(qarg1):
-            # [[q[0], q[1]], [r[0], r[1]]] -> [q[0], r[0]]
-            #                              -> [q[1], r[1]]
-            for arg0, arg1 in zip(qarg0, qarg1):
-                yield [arg0, arg1], []
-        elif len(qarg0) == 1:
-            # [[q[0]], [r[0], r[1]]] -> [q[0], r[0]]
-            #                        -> [q[0], r[1]]
-            for arg1 in qarg1:
-                yield [qarg0[0], arg1], []
-        elif len(qarg1) == 1:
-            # [[q[0], q[1]], [r[0]]] -> [q[0], r[0]]
-            #                        -> [q[1], r[0]]
-            for arg0 in qarg0:
-                yield [arg0, qarg1[0]], []
-        else:
-            raise CircuitError(
-                f"Not sure how to combine these two-qubit arguments:\n {qarg0}\n {qarg1}"
-            )
-
-    @staticmethod
-    def _broadcast_3_or_more_args(qargs: List) -> List:
-        if all(len(qarg) == len(qargs[0]) for qarg in qargs):
-            for arg in zip(*qargs):
-                yield list(arg), []
-        else:
-            raise CircuitError("Not sure how to combine these qubit arguments:\n %s\n" % qargs)
-
-    def broadcast_arguments(self, qargs: List, cargs: List) -> Tuple[List, List]:
-        """Validation and handling of the arguments and its relationship.
-
-        For example, ``cx([q[0],q[1]], q[2])`` means ``cx(q[0], q[2]); cx(q[1], q[2])``. This
-        method yields the arguments in the right grouping. In the given example::
-
-            in: [[q[0],q[1]], q[2]],[]
-            outs: [q[0], q[2]], []
-                  [q[1], q[2]], []
-
-        The general broadcasting rules are:
-
-            * If len(qargs) == 1::
-
-                [q[0], q[1]] -> [q[0]],[q[1]]
-
-            * If len(qargs) == 2::
-
-                [[q[0], q[1]], [r[0], r[1]]] -> [q[0], r[0]], [q[1], r[1]]
-                [[q[0]], [r[0], r[1]]]       -> [q[0], r[0]], [q[0], r[1]]
-                [[q[0], q[1]], [r[0]]]       -> [q[0], r[0]], [q[1], r[0]]
-
-            * If len(qargs) >= 3::
-
-                [q[0], q[1]], [r[0], r[1]],  ...] -> [q[0], r[0], ...], [q[1], r[1], ...]
-
-        Args:
-            qargs: List of quantum bit arguments.
-            cargs: List of classical bit arguments.
-
-        Returns:
-            A tuple with single arguments.
-
-        Raises:
-            CircuitError: If the input is not valid. For example, the number of
-                arguments does not match the gate expectation.
-        """
-        if len(qargs) != self.num_qubits or cargs:
-            raise CircuitError(
-                f"The amount of qubit({len(qargs)})/clbit({len(cargs)}) arguments does"
-                f" not match the gate expectation ({self.num_qubits})."
-            )
-
-        if any(not qarg for qarg in qargs):
-            raise CircuitError("One or more of the arguments are empty")
-
-        if len(qargs) == 1:
-            return Gate._broadcast_single_argument(qargs[0])
-        elif len(qargs) == 2:
-            return Gate._broadcast_2_arguments(qargs[0], qargs[1])
-        elif len(qargs) >= 3:
-            return Gate._broadcast_3_or_more_args(qargs)
-        else:
-            raise CircuitError("This gate cannot handle %i arguments" % len(qargs))
 
     def validate_parameter(self, parameter):
         """Gate parameters should be int, float, or ParameterExpression"""
