+++ qiskit/circuit/measure.py
@@ -19,27 +19,15 @@ import warnings
 from qiskit.circuit.instruction import Instruction
 from qiskit.circuit.operation import Operation
 from qiskit.circuit.exceptions import CircuitError
+from .argumentsbroadcaster import ArgumentsBroadcasterMeasure
 
 
-class Measure(Instruction, Operation):
+class Measure(ArgumentsBroadcasterMeasure, Instruction, Operation):
     """Quantum measurement in the computational basis."""
 
     def __init__(self):
         """Create new measurement instruction."""
         super().__init__("measure", 1, 1, [])
-
-    def broadcast_arguments(self, qargs, cargs):
-        qarg = qargs[0]
-        carg = cargs[0]
-
-        if len(carg) == len(qarg):
-            for qarg, carg in zip(qarg, carg):
-                yield [qarg], [carg]
-        elif len(qarg) == 1 and carg:
-            for each_carg in carg:
-                yield qarg, [each_carg]
-        else:
-            raise CircuitError("register size error")
 
 
 def measure(circuit, qubit, clbit):
